refactor: Moved frame resolution logic to dataset_loader_utils
